+++ src/app/brokers/page.tsx
@@ -1,0 +1,21 @@
+import { useTranslations } from 'next-intl';
+import * as React from 'react';
+
+
+export default function AlpacaDashboard() {
+
+  const t = useTranslations('Alpaca');
+  return (
+    <div className="content-container w-full pt-[10%] text-slate-800 pl-10">
+        <h1 className='font-bold text-5xl'>Broker</h1>
+        <h2 className='font-semibold text-2xl mt-6'>Erklärungen zu Brokern</h2>
+        <p>Wählen Sie einen Broker</p>
+        <p>Rufen Sie die Einstellungen auf</p>
+        <p>Tragen Sie Ihre Login daten für die Broker API ein</p>
+        <p>Speichern Sie die Einstellungen</p>
+
+     
+    </div>
+
+  );
+}