+++ src/components/alpaca/history-charts/AlpacaHistoryChart.tsx
@@ -9,8 +9,6 @@ import { ChartParameters } from '@/model/ChartParameters';
 
 const fetcher = (url: string) => fetch(url).then((res) => res.json());
 
-
-
 const AlpacaHistoryChart = () => {
 
     const [params, setParameters] = useState<ChartParameters>({