+++ src/model/Asset.ts
@@ -1,4 +1,0 @@
-export interface Asset {
-    symbol: string;
-    name: string;
-}